pyproject.toml
[project]
name = "pml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arviz>=0.22.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pyro-ppl>=1.9.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
]
